+++ R/utils.R
@@ -40,7 +40,11 @@ get_region <- function(x, ...) {
 }
 
 get_region.default <- function(x, ...) {
-    get_location(bucket = x)
+    out <- get_location(bucket = x)
+    if (!is.character(out)) {
+        stop("Cannot detect correct region from bucket")
+    }
+    out
 }
 
 get_region.s3_bucket <- function(x, ...) {